+++ backend/src/main/java/com/popcorn/demo/infrastructure/config/InfrastructureConfig.java
@@ -1,7 +1,8 @@
-package com.ttalkak.demo.infrastructure.config;
+package com.popcorn.demo.infrastructure.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 /**
@@ -9,10 +10,11 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
  */
 @Configuration
 @ComponentScan(basePackages = {
-    "com.ttalkak.demo.infrastructure.persistence",
-    "com.ttalkak.demo.infrastructure.external"
+    "com.popcorn.demo.infrastructure.persistence",
+    "com.popcorn.demo.infrastructure.external"
 })
-@EnableJpaRepositories(basePackages = "com.ttalkak.demo.infrastructure.persistence.repository")
+@EnableJpaAuditing
+@EnableJpaRepositories(basePackages = "com.popcorn.demo.infrastructure.persistence.repository")
 public class InfrastructureConfig {
     // 필요시 추가 Bean 설정
 }
