+++ backend/src/test/java/com/popcorn/demo/DemoApplicationTests.java
@@ -1,9 +1,11 @@
-package com.ttalkak.demo;
+package com.popcorn.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class DemoApplicationTests {
 
 	@Test
